chore: Removed a binding to Tab in cmp.lua
Have had the problem where when I type something and I get the
suggestions pop up but just want to insert an Tab. The first item in the
suggestions option get selected...

With this change I have commented out the section with this binding :
['<Tab>'] = cmp.mapping(function(fallback)

First test seems to solve the problem, but need to use neovim some to
see if there are bugs introduced by removing it. So the removed code is
kept in the file commented out.

modified:   lua/custom/plugins/cmp.lua