feat: More bindings in my_keymaps.lua
Added back J and K to move visual selected blocs up and down. the
MINI.move also have this but uses Alt + J and K. I prefer only J and K

Added a rebind of J (joining line bellow, behind current line) so that
cursor stay where it is when joining. Normal is for the cursor to jump
back behind the line when joining.

modified:   my_keymaps.lua